\name{apollo13-package}
\alias{apollo13-package}
\alias{apollo13}
\docType{package}
\title{
  Houston, We Have A Problem
}
\description{
  Some people write packages with too many underscores, and too many
  imports shadowing other packages.  This package addresses both
  problems by selectively importing just a few key funtions and
  carefully renaming them. To camelCase, naturally.
}
\details{
  This mostly uses facilities now available via the NAMESPACE mechanism.
}
\author{
  Dirk Eddelbuettel

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\keyword{package}
