% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aliases.R
\name{buildPackage}
\alias{buildPackage}
\title{buildPackage}
\usage{
buildPackage(...)
}
\arguments{
\item{...}{passed on to \code{build}}
}
\description{
buildPackage
}
\seealso{
\code{\link[devtools]{build}}
}

