% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintainers.R
\name{get_maintainers}
\alias{get_maintainers}
\title{Maintainers}
\usage{
get_maintainers(universe)
}
\arguments{
\item{universe}{universe to query}
}
\value{
A character vector.
}
\description{
Maintainers
}
\examples{
get_maintainers("ggseg")
}
