% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_packages}
\alias{get_packages}
\title{Packages}
\usage{
get_packages(universe)
}
\arguments{
\item{universe}{universe to query}
}
\value{
A character vector.
}
\description{
Packages
}
\examples{
get_packages("ropensci")
}
