% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aliases.R
\name{installGithub}
\alias{installGithub}
\title{installGithub}
\usage{
installGithub(...)
}
\arguments{
\item{...}{passed on to \code{install_github}}
}
\description{
installGithub
}
\seealso{
\code{\link[devtools]{install_github}}
}

